AFK Pause 0.7.1
