--- AFKTimePause/Methods.cs
@@ -1,19 +1,0 @@
-﻿using HarmonyLib;
-using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using Microsoft.Xna.Framework.Input;
-using Newtonsoft.Json;
-using StardewValley;
-using StardewValley.Menus;
-using StardewValley.Objects;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Xml.Serialization;
-
-namespace AFKTimePause
-{
-    public partial class ModEntry
-    {
-    }
-}
