--- AFKTimePause/i18n/ru.json
@@ -1,10 +1,0 @@
-{
-  // GMCM
-  "GMCM_Option_ModEnabled_Name": "Мод включён?",
-  "GMCM_Option_FreezeGame_Name": "Заморозить игру",
-  "GMCM_Option_FreezeGame_Tooltip": "По умолчанию, этот мод замораживает игру: это схоже с открытием инвентаря. Если эта опция будет включена, персонаж продолжит перемещение даже в AFK, но игровые часы будут остановлены.",
-  "GMCM_Option_TicksTilAFK_Name": "Тиков до AFK",
-  "GMCM_Option_ShowAFKText_Name": "Отображать текст в AFK",
-  "GMCM_Option_AFKText_Name": "Текст в AFK",
-  "GMCM_Option_WakeOnMouseMove_Name": "Пробуждение от мыши"
-}
