--- AFKTimePause/manifest.json
@@ -1,22 +1,0 @@
-﻿{
-  "Name": "AFK Time Pause",
-  "Author": "aedenthorn",
-  "Version": "0.7.0",
-  "Description": "AFK Time Pause.",
-  "UniqueID": "aedenthorn.AFKTimePause",
-  "EntryDll": "AFKTimePause.dll",
-  "MinimumApiVersion": "3.15.0",
-  "ModUpdater": {
-    "Repository": "StardewValleyMods",
-    "User": "aedenthorn",
-    "Directory": "_releases",
-    "ModFolder": "AFKTimePause"
-  },
-  "Dependencies": [
-    {
-      "UniqueID": "Platonymous.ModUpdater",
-      "IsRequired": false
-    }
-  ],
-  "UpdateKeys": [ "Nexus:14228" ]
-}
